Sqlite: Return rowid values as long instead of int (bgo#696520)
When the command type is "execute", the result is the rowid of the last
inserted row (connection.LastInsertRowId), which is a long.

We now return the value as a long, instead of converting it to an int
ourselves, and throwing an exception when the value is above
Int32.MaxValue.

In SQLite, a column defined as INTEGER PRIMARY KEY is used as the rowid,
so it should now be handled as a long.

As a consequence, for a class using SqliteModelProvider, the member
marked as the primary key must be a long. Previously, it was assumed to
be assignable from an int, so we add a specific runtime check for that.

+++ Hyena.Data.Sqlite/Hyena.Data.Sqlite/Tests/DbBoundType.cs
@@ -81,7 +81,7 @@ namespace Hyena.Data.Sqlite.Tests
     internal class DbBoundType
     {
         [DatabaseColumn ("PrimaryKey", Constraints = DatabaseColumnConstraints.PrimaryKey)]
-        public int PrimaryKey;
+        public long PrimaryKey;
 
         [DatabaseColumn ("PublicIntField")]
         public int PublicIntField;
